Move shared under app
