+++ components/form-builder/app/Settings.tsx
@@ -2,13 +2,13 @@ import React, { useState } from "react";
 import { useTranslation } from "next-i18next";
 import { useRouter } from "next/router";
 import { useTemplateStore } from "../store/useTemplateStore";
-import { Button } from "../shared/Button";
-import { Input } from "../shared/Input";
+import { Button } from "./shared/Button";
+import { Input } from "./shared/Input";
 import { useSession } from "next-auth/react";
 import { useDeleteForm } from "../hooks/useDelete";
 import Markdown from "markdown-to-jsx";
-import { useDialogRef, Dialog } from "../shared/Dialog";
-import { ConfirmFormDeleteDialog } from "../shared/ConfirmFormDeleteDialog";
+import { useDialogRef, Dialog } from "./shared/Dialog";
+import { ConfirmFormDeleteDialog } from "./shared/ConfirmFormDeleteDialog";
 
 const FormDeleted = () => {
   const { t } = useTranslation("form-builder");
