+++ components/form-builder/app/edit/Modal.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import { useTranslation } from "next-i18next";
 import PropTypes from "prop-types";
 
-import { Button } from "../../shared";
+import { Button } from "../shared";
 import { Close } from "../../icons";
 import { CDSHTMLDialogElement } from "../../types";
 import { useModalStore } from "../../store";
