+++ components/form-builder/app/edit/ModalForm.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from "next-i18next";
 import { FormElementTypes, ElementProperties } from "@lib/types";
 
 import { FormElementWithIndex, LocalizedElementProperties } from "../../types";
-import { Checkbox, Input, TextArea } from "../../shared";
+import { Checkbox, Input, TextArea } from "../shared";
 import { useTemplateStore } from "../../store";
 
 const FormLabel = styled.label`
