+++ components/form-builder/app/edit/elements/Options.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from "next-i18next";
 import { useTemplateStore } from "../../../store/useTemplateStore";
 import { Option } from "./Option";
 import { BulkAdd } from "./BulkAdd";
-import { Button } from "../../../shared/Button";
+import { Button } from "../../shared/Button";
 import { FormElementWithIndex } from "../../../types";
 
 const AddButton = ({ index, onClick }: { index: number; onClick: (index: number) => void }) => {
